backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
websockets==13.0

# Database
sqlalchemy==2.0.35
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching & Queue
redis==5.0.7
celery==5.3.4
flower==2.0.1

# API Integrations
openai==1.40.0
anthropic==0.34.2  # Claude API SDK
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# Utils
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
pendulum==3.0.0
shortuuid==1.0.13
playwright==1.40.0
dnspython==2.4.2

# Content & Text Analysis
textstat==0.7.3
nltk==3.8.1
language-tool-python==2.7.1

# Monitoring
sentry-sdk[fastapi]==2.12.0
prometheus-client==0.20.0
structlog==24.4.0
posthog==3.3.4

# Testing
pytest==8.3.2
pytest-asyncio==0.23.8
pytest-cov==5.0.0
pytest-httpx==0.30.0

# Development
black==24.8.0
ruff==0.6.2
mypy==1.11.1
pre-commit==3.8.0

# Google Analytics Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-analytics-data==0.18.0
cryptography==41.0.7